streamlit_app/requirements.txt
streamlit==1.36.0
pandas==1.5.3
numpy==1.24.4
scikit-learn==1.3.2
joblib==1.3.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
dill